requirements.txt
asgiref==3.5.0
Django==4.0.2
django-js-asset==2.0.0
django-mptt==0.13.4
django-treewidget==0.4.0
djangorestframework==3.13.1
numexpr==2.8.1
numpy==1.22.2
packaging==21.3
pyparsing==3.0.7
pytz==2021.3
sqlparse==0.4.2
